requirements/prod.txt
# Install everything that needs to be compiled
-r compiled.txt
PIL
gunicorn

# Django stuff
Django>=1.4

# Templates
jingo
jingo-minify
Markdown
#Markdown==2.1.1
-e git://github.com/hcarvalhoalves/django-taggit.git#egg=django-taggit
django-intellipages
#-e git://github.com/grengojbo/django-flatpages-plus.git#egg=django-flatpages-plus
django-widget-tweaks

# Translate
django-modeltranslation
django-modeltranslation-wrapper

# Security
bleach
django-session-csrf
#django-registration
django-guardian
#cef

# Profile
-e git://github.com/grengojbo/django-userena.git#egg=userena

# Logging
#raven

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower
pytils

# Ajax
django-dajaxice
django-dajax

# Caching
python-memcached

# Admin
django-debug-toolbar
#django-debug-toolbar==0.8.5
-e git+git://github.com/sehmaschine/django-grappelli.git#egg=django-grappelli
#django-grappelli>=2.4.2
-e git+git://github.com/h3/grappelli-fit.git#egg=grappellifit
-e git://github.com/h3/django-seoutils.git#egg=seoutils
-e git://github.com/sehmaschine/django-filebrowser.git#egg=django-filebrowser
# There are useful panels for debug-toolbar, but have not quite been updated yet for 1.4
#-e git://github.com/playfire/django-debug-toolbar-user-panel#egg=django_debug_toolbar_user_panel-dev
#-e git://github.com/ross/memcache-debug-panel#egg=memcache_toolbar-dev
#-e git+git://github.com/django-extensions/django-extensions.git#egg=django_extensions
django_extensions

# HelpDesk
-e git://github.com/grengojbo/django-knowledge.git#egg=knowledge

# Migrations
South
#South==0.7.3

# Various tidbits
commonware
-e git://github.com/mozilla/nuggets.git#egg=nuggets

